+++ ios/THEOplayerRCTMainEventHandler.swift
@@ -30,6 +30,7 @@ public class THEOplayerRCTMainEventHandler {
     var onNativeRateChange: RCTDirectEventBlock?
     var onNativeWaiting: RCTDirectEventBlock?
     var onNativeCanPlay: RCTDirectEventBlock?
+    var onNativeContentProtectionError: RCTDirectEventBlock?
     var onNativeDimensionChange: RCTDirectEventBlock?
     var onNativeVideoResize: RCTDirectEventBlock?
     
@@ -54,6 +55,7 @@ public class THEOplayerRCTMainEventHandler {
     private var rateChangeListener: EventListener?
     private var waitingListener: EventListener?
     private var canPlayListener: EventListener?
+    private var contentProtectionErrorListener: EventListener?
     private var videoResizeListener: EventListener?
     
     // MARK: player observer
@@ -326,6 +328,25 @@ public class THEOplayerRCTMainEventHandler {
         }
         if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] CanPlay listener attached to THEOplayer") }
         
+        // CONTENT_PROTECTION_ERROR
+        self.contentProtectionErrorListener = player.addEventListener(type: PlayerEventTypes.CONTENT_PROTECTION_ERROR) { [weak self] event in
+            if DEBUG_THEOPLAYER_EVENTS { PrintUtils.printLog(logText: "[NATIVE] Received CONTENT_PROTECTION_ERROR event from THEOplayer") }
+            if let forwardedContentProtectionErrorEvent = self?.onNativeContentProtectionError,
+               let errorObject = event.errorObject {
+                let errorCodeString = String(errorObject.code.rawValue)
+                let errorCodeMessage = errorObject.message
+                forwardedContentProtectionErrorEvent(
+                    [
+                        "error": [
+                            "errorCode": errorCodeString,
+                            "errorMessage": errorCodeMessage
+                        ]
+                    ]
+                )
+            }
+        }
+        if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] ContentProtectionError listener attached to THEOplayer") }
+        
         // RESIZE
         self.videoResizeListener = player.addEventListener(type: PlayerEventTypes.RESIZE) { [weak self, weak player] event in
             if DEBUG_THEOPLAYER_EVENTS { PrintUtils.printLog(logText: "[NATIVE] Received RESIZE event from THEOplayer") }
@@ -486,6 +507,12 @@ public class THEOplayerRCTMainEventHandler {
         self.dimensionChangeObserver?.invalidate()
         self.dimensionChangeObserver = nil
       
+        // CONTENT_PROTECTION_ERROR
+        if let contentProtectionErrorListener = self.contentProtectionErrorListener {
+            player.removeEventListener(type: PlayerEventTypes.CONTENT_PROTECTION_ERROR, listener: contentProtectionErrorListener)
+            if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] ContentProtectionError listener dettached from THEOplayer") }
+        }
+        
         // RESIZE
         if let videoResizeListener = self.videoResizeListener {
           player.removeEventListener(type: PlayerEventTypes.RESIZE, listener: videoResizeListener)